// completedProject/src/models/species.model.ts
export class Species {
    name:             string;
    classification:   string;
    designation:      string;
    average_height:   string;
    average_lifespan: string;
    eye_colors:       string;
    hair_colors:      string;
    skin_colors:      string;
    language:         string;
    created:          string;
    edited:           string;
}